fix(datapack): share source snapshot URI policy
