+++ tools/datapack/build-datapack.mjs
@@ -6,6 +6,7 @@ import { DatabaseSync } from "node:sqlite";
 import path from "node:path";
 import { tmpdir } from "node:os";
 import { usesLocalPlaceholderHost } from "./production-url-policy.mjs";
+import { requiredCredentialFreeObjectUri } from "./source-snapshot-policy.mjs";
 
 const root = path.resolve(import.meta.dirname, "../..");
 const productionMinimumTableRowNames = [
@@ -508,28 +509,6 @@ function validatePackUrlMatchesStagedPath(packUrl, pack, label) {
   if (!url.pathname.endsWith(expectedPathSuffix) || url.search !== "" || url.hash !== "") {
     throw new Error(`${label} absolute HTTPS URL path must end with ${stagedPackPath(pack)}`);
   }
-}
-
-function requiredCredentialFreeObjectUri(value, label) {
-  const uri = requiredString(value, label);
-  let parsed;
-  try {
-    parsed = new URL(uri);
-  } catch {
-    throw new Error(`${label} must be a credential-free object storage URI`);
-  }
-  if (!["s3:", "oci:"].includes(parsed.protocol)
-    || parsed.username !== ""
-    || parsed.password !== ""
-    || parsed.search !== ""
-    || parsed.hash !== ""
-    || parsed.hostname === ""
-    || parsed.pathname === ""
-    || parsed.pathname === "/"
-    || uri.includes("@")) {
-    throw new Error(`${label} must be a credential-free object storage URI`);
-  }
-  return uri;
 }
 
 function stagedPackPath(pack) {
