+++ tools/datapack/build-source-snapshot.mjs
@@ -2,6 +2,7 @@
 import { createHash } from "node:crypto";
 import { mkdir, readFile, writeFile } from "node:fs/promises";
 import path from "node:path";
+import { requiredCredentialFreeObjectUri } from "./source-snapshot-policy.mjs";
 
 async function main() {
   const args = parseArgs(process.argv.slice(2));
@@ -139,28 +140,6 @@ function validateSnapshot(snapshot) {
 async function writeFileWithParents(filePath, body) {
   await mkdir(path.dirname(path.resolve(filePath)), { recursive: true });
   await writeFile(filePath, body);
-}
-
-function requiredCredentialFreeObjectUri(value, label) {
-  const uri = requiredText(value, label);
-  let parsed;
-  try {
-    parsed = new URL(uri);
-  } catch {
-    throw new Error(`${label} must be a credential-free object storage URI`);
-  }
-  if (!["s3:", "oci:"].includes(parsed.protocol)
-    || parsed.username !== ""
-    || parsed.password !== ""
-    || parsed.search !== ""
-    || parsed.hash !== ""
-    || parsed.hostname === ""
-    || parsed.pathname === ""
-    || parsed.pathname === "/"
-    || uri.includes("@")) {
-    throw new Error(`${label} must be a credential-free object storage URI`);
-  }
-  return uri;
 }
 
 function requiredDate(value, label) {
